messaging_app/requirements.txt
aiosqlite==0.21.0
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.8
django-seed==0.3.1
djangorestframework==3.16.1
drf-nested-routers==0.95.0
Faker==38.0.0
idna==3.11
parameterized==0.9.0
psycopg2==2.9.11
requests==2.32.5
sqlparse==0.5.3
toposort==1.10
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
